// src/permissions/entities/permissions.enum.ts
export enum Permission {
  CREATE_USER = 'create_user',
  UPDATE_USER = 'update_user',
  VIEW_USER = 'view_user',
  DELETE_USER = 'delete_user',
  CREATE_PETITION = 'create_petition',
  UPDATE_PETITION = 'update_petition',
  VIEW_PETITION = 'view_petition',
  DELETE_PETITION = 'delete_petition',
  SEND_REQUEST = 'send_request',
  ACCEPT_REQUEST = 'accept_request',
  UPLOAD_SIGNED_PETITIONS = 'upload_signed_petitions',
  VIEW_SIGNED_PETITIONS = 'view_signed_petitions',
  VALIDATE_SIGNED_PETITIONS = 'validate_signed_petitions'
}